fix(ci): split analysis API tests and add journey manifests dir
Split test_analysis_api.py under the 500-line LOC guard by moving
trace-matrix export and shortest-path tests into dedicated modules.
Add docs/journeys/manifests/.gitkeep so journey-verify passes.
