+++ tests/unit/api/test_analysis_api.py
@@ -1,7 +1,4 @@
-"""Comprehensive tests for Analysis API endpoints.
-
-Tests impact analysis, cycle detection, and shortest path finding.
-"""
+"""Tests for Analysis API endpoints (impact analysis and cycle detection)."""
 
 from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
@@ -307,269 +304,6 @@ class TestCycleDetection:
         response = client.get("/api/v1/analysis/cycles/")
         # Should fail due to missing project_id in path
         assert response.status_code in {404, 422}
-
-
-class TestShortestPath:
-    """Test shortest path finding endpoint."""
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_simple(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path with simple dependency chain."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        mock_result = MagicMock(
-            exists=True,
-            distance=2,
-            path=["item1", "item2", "item3"],
-            link_types=["depends_on", "depends_on"],
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item3"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["exists"] is True
-            assert data["distance"] == COUNT_TWO
-            assert len(data["path"]) == COUNT_THREE
-            assert len(data["link_types"]) == COUNT_TWO
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_long_chain(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path with long dependency chain."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        path = [f"item{i}" for i in range(1, 11)]  # 10 items
-        link_types = ["depends_on"] * 9  # 9 links
-
-        mock_result = MagicMock(
-            exists=True,
-            distance=9,
-            path=path,
-            link_types=link_types,
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item10"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["distance"] == 9
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_not_found(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path when no path exists."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        mock_result = MagicMock(
-            exists=False,
-            distance=None,
-            path=[],
-            link_types=[],
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item999"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["exists"] is False
-            assert data["distance"] is None
-            assert data["path"] == []
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_direct_link(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path with direct link between items."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        mock_result = MagicMock(
-            exists=True,
-            distance=1,
-            path=["item1", "item2"],
-            link_types=["depends_on"],
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item2"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["exists"] is True
-            assert data["distance"] == 1
-            assert len(data["path"]) == COUNT_TWO
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_missing_source_id(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path requires source_id parameter."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_db.return_value = AsyncMock()
-
-        response = client.get("/api/v1/analysis/shortest-path?project_id=proj1&target_id=item2")
-        # Missing source_id should cause an error
-        assert response.status_code in {422, 500}
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_missing_target_id(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path requires target_id parameter."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_db.return_value = AsyncMock()
-
-        response = client.get("/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1")
-        # Missing target_id should cause an error
-        assert response.status_code in {422, 500}
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_missing_project_id(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path requires project_id parameter."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_db.return_value = AsyncMock()
-
-        response = client.get("/api/v1/analysis/shortest-path?source_id=item1&target_id=item2")
-        # Missing project_id should cause an error
-        assert response.status_code in {422, 500}
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_multiple_link_types(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path with various link types."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        mock_result = MagicMock(
-            exists=True,
-            distance=3,
-            path=["item1", "item2", "item3", "item4"],
-            link_types=["depends_on", "related_to", "blocks"],
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item4"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["link_types"] == ["depends_on", "related_to", "blocks"]
-
-    @patch("tracertm.api.main.get_db")
-    @patch("tracertm.api.main.auth_guard")
-    def test_shortest_path_self_reference(self, mock_auth: Any, mock_db: Any) -> None:
-        """Test shortest path from item to itself."""
-        mock_auth.return_value = {"role": "user", "sub": "user123"}
-        mock_session = AsyncMock()
-        mock_db.return_value = mock_session
-
-        mock_result = MagicMock(
-            exists=True,
-            distance=0,
-            path=["item1"],
-            link_types=[],
-        )
-
-        with patch(
-            "tracertm.api.main.shortest_path_service.ShortestPathService"
-        ) as mock_service_class:
-            mock_service = MagicMock()
-            mock_service.find_shortest_path.return_value = mock_result
-            mock_service_class.return_value = mock_service
-
-            response = client.get(
-                "/api/v1/analysis/shortest-path?project_id=proj1&source_id=item1&target_id=item1"
-            )
-            assert response.status_code == HTTP_OK
-            data = response.json()
-            assert data["distance"] == 0
-
-
-class TestTraceMatrixExport:
-    """Test traceability matrix CSV export endpoint."""
-
-    @patch("tracertm.api.routers.analysis.ensure_project_access")
-    @patch("tracertm.api.routers.analysis.enforce_rate_limit")
-    def test_export_trace_matrix_csv(self, _mock_rate: Any, _mock_access: Any) -> None:
-        """CSV export returns attachment with matrix content."""
-        mock_session = AsyncMock()
-
-        async def override_db() -> AsyncMock:
-            return mock_session
-
-        def override_auth() -> dict[str, str]:
-            return {"role": "user", "sub": "user123"}
-
-        app.dependency_overrides[get_db] = override_db
-        app.dependency_overrides[auth_guard] = override_auth
-
-        mock_matrix = MagicMock()
-
-        try:
-            with patch(
-                "tracertm.api.routers.analysis.traceability_matrix_service.TraceabilityMatrixService",
-            ) as mock_service_class:
-                mock_service = MagicMock()
-                mock_service.generate_matrix = AsyncMock(return_value=mock_matrix)
-                mock_service.export_matrix_csv = AsyncMock(
-                    return_value='"Source","Feature A"\n"Req 1","traces_to"',
-                )
-                mock_service_class.return_value = mock_service
-
-                response = client.get("/api/v1/analysis/trace-matrix/export?project_id=proj1")
-                assert response.status_code == HTTP_OK
-                assert response.headers["content-type"].startswith("text/csv")
-                disposition = response.headers.get("content-disposition", "")
-                assert "attachment" in disposition
-                assert "tracera-matrix" in disposition
-                assert "Req 1" in response.text
-        finally:
-            app.dependency_overrides.clear()
 
 
 if __name__ == "__main__":
